[buffer] Basic non-concurrent implementation

+++ source/laplace/execution.h
@@ -4,6 +4,7 @@
 #ifndef LAPLACE_EXECUTION_H
 #define LAPLACE_EXECUTION_H
 
+#include <cstddef>
 #include <shared_mutex>
 #include <thread>
 #include <vector>
@@ -11,24 +12,23 @@
 namespace laplace {
   class execution {
   public:
-    static const signed long long default_thread_count;
-    static const signed long long overthreading_limit;
+    static const ptrdiff_t default_thread_count;
+    static const ptrdiff_t overthreading_limit;
 
     execution() noexcept;
     execution(execution const &exe) noexcept;
     execution(execution &&exe) noexcept;
     ~execution() noexcept;
 
-    auto operator =(execution const &exe) noexcept -> execution &;
-    auto operator =(execution &&exe) noexcept -> execution &;
+    auto operator=(execution const &exe) noexcept -> execution &;
+    auto operator=(execution &&exe) noexcept -> execution &;
 
     [[nodiscard]] auto is_error() const noexcept -> bool;
 
-    [[nodiscard]] auto get_thread_count() const noexcept
-        -> signed long long;
+    [[nodiscard]] auto get_thread_count() const noexcept -> ptrdiff_t;
 
     [[nodiscard]] auto set_thread_count(
-        signed long long thread_count) const noexcept -> execution;
+        ptrdiff_t thread_count) const noexcept -> execution;
 
   private:
     [[nodiscard]] static auto _error() noexcept -> execution;
